Move CPU stats to a separate hwstats package. Fixes #659.
