+++ .changeset/polite-hounds-tickle.md
@@ -1,0 +1,5 @@
+---
+"github.com/livekit/protocol": minor
+---
+
+Moved CPU stats to a separate hwstats package, removing cgo dependency.
